pyproject.toml
[project]
# You can change the name and version, but they are required
name = "prompt-market"
version = "0.1.0"
description = "Python project managed by uv"

# The core list of all necessary external packages
dependencies = [
    "datasets",
    "python-dotenv",
    "numpy",
    "pandas",
    "scipy",
    "matplotlib",
    "openai",
    "groq",
    "google-genai",
]